requirements.txt
# Function dependencies, for example:
# package>=version
recipe-scrapers
google-cloud-firestore
google-cloud-storage
pillow